requirements.txt
pandas
pillow
torch
torchvision
torch-directml